+++ src/libs/test-libs/abilib/abiTest/options.cpp
@@ -17,7 +17,7 @@ CParams params[] = {
     CParams("~mode", "Mode of operation. One or more of 'encoding' or 'generation'.\n"),
     CParams("",      "Simple program to illustrate how to encode function and event signatures.\n"),
 };
-uint32_t nParams = sizeof(params) / sizeof(CParams);
+size_t nParams = sizeof(params) / sizeof(CParams);
 
 //---------------------------------------------------------------------------------------------------
 bool COptions::parseArguments(string_q& command) {
