+++ src/libs/test-libs/etherlib/dataUpgrade/newblock.h
@@ -50,7 +50,7 @@ public:
 
     DECLARE_NODE(CNewBlock);
 
-    const CBaseNode *getObjectAt(const string_q& fieldName, uint32_t index) const override;
+    const CBaseNode *getObjectAt(const string_q& fieldName, size_t index) const override;
 
     // EXISTING_CODE
     CNewBlock(const CBlock& block);
