+++ src/libs/utillib/namevalue.cpp
@@ -116,7 +116,7 @@ void CNameValue::registerClass(void) {
     if (been_here) return;
     been_here = true;
 
-    uint32_t fieldNum = 1000;
+    size_t fieldNum = 1000;
     ADD_FIELD(CNameValue, "schema",  T_NUMBER, ++fieldNum);
     ADD_FIELD(CNameValue, "deleted", T_BOOL,  ++fieldNum);
     ADD_FIELD(CNameValue, "showing", T_BOOL,  ++fieldNum);
