+++ src/tools/getLogs/getLogs.cpp
@@ -63,7 +63,7 @@ bool visitTransaction(CTransaction& trans, void *data) {
     }
 
 	cout << "[";
-	for (uint32_t i = 0 ; i < trans.receipt.logs.size() ; i++) {
+	for (size_t i = 0 ; i < trans.receipt.logs.size() ; i++) {
 		trans.receipt.logs[i].doExport(cout);
 		cout << (i < trans.receipt.logs.size()-1 ? ",\n" : "\n");
 	}
